requirements.txt
-i https://pypi.org/simple
alembic==1.14.1; python_version >= '3.8'
aniso8601==10.0.1
attrs==25.3.0; python_version >= '3.8'
backports.entry-points-selectable==1.3.0; python_version >= '3.8'
bcrypt==4.3.0; python_version >= '3.8'
blinker==1.8.2; python_version >= '3.8'
click==8.1.8; python_version >= '3.7'
faker==35.2.2; python_version >= '3.8'
flask==3.0.3; python_version >= '3.8'
flask-bcrypt==1.0.1
flask-cors==5.0.0
flask-jwt-extended==4.6.0; python_version >= '3.7' and python_version < '4'
flask-migrate==4.1.0; python_version >= '3.6'
flask-restx==1.3.0
flask-sqlalchemy==3.1.1; python_version >= '3.8'
greenlet==3.1.1; python_version >= '3.7'
gunicorn==23.0.0; python_version >= '3.7'
honcho==2.0.0
importlib-metadata==8.5.0; python_version >= '3.8'
importlib-resources==6.4.5; python_version >= '3.8'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
mako==1.3.10; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
packaging==25.0; python_version >= '3.8'
pkgutil-resolve-name==1.3.10; python_version >= '3.6'
psycopg2-binary==2.9.10; python_version >= '3.8'
pyjwt==2.9.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2025.2
referencing==0.35.1; python_version >= '3.8'
rpds-py==0.20.1; python_version >= '3.8'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlalchemy==2.0.41; python_version >= '3.7'
sqlalchemy-serializer==1.4.12
typing-extensions==4.13.2; python_version >= '3.8'
werkzeug==3.0.6; python_version >= '3.8'
zipp==3.20.2; python_version >= '3.8'
